refactor(world): simplify pass on object-material ownership
Post-review cleanups (behaviour-preserving, 192 tests + render pin green):
- Gate HitRecord::new behind #[cfg(test)] — production builds records only via
  from_geo (the World is the sole place material binds to a hit), so the
  invariant is now a compile error, not a doc comment.
- Delete dead ObjData::into_mesh / into_triangles (no callers; the ownership
  change had been maintaining their material params for nothing) and drop the
  imports they freed.
- Drop a redundant `as f32` cast in IntersectGroup::intersect (GeoHit::t is f32).

Kept World::intersect's accumulate-then-attach form: it localizes material
attachment to one line — the exact site the future material-slot index edits —
and is marginally cheaper than attaching inside the loop.

+++ src/group.rs
@@ -55,7 +55,7 @@ impl Intersect for IntersectGroup {
 
         for object in &self.objects {
             if let Some(hit_record) = object.intersect(ray, &Interval::new(ray_t.min, closest_t)) {
-                closest_t = hit_record.t as f32;
+                closest_t = hit_record.t;
                 closest_hit = Some(hit_record);
             }
         }
